Fixes non-integer indexing issue

+++ perceptual/filterbank.py
@@ -149,8 +149,8 @@ class Steerable:
 			# ============== Lowpass component are upsampled and convoluted ============
 			dims = np.array(coeff[0][0].shape)
 			
-			lostart = np.ceil((dims+0.5)/2) - np.ceil((np.ceil((dims-0.5)/2)+0.5)/2) 
-			loend = lostart + np.ceil((dims-0.5)/2) 
+			lostart = (np.ceil((dims+0.5)/2) - np.ceil((np.ceil((dims-0.5)/2)+0.5)/2)).astype(np.int32)
+			loend = lostart + np.ceil((dims-0.5)/2).astype(np.int32) 
 
 			nlog_rad = log_rad[lostart[0]:loend[0], lostart[1]:loend[1]]
 			nangle = angle[lostart[0]:loend[0], lostart[1]:loend[1]]
